requirements.txt
pandas
numpy
matplotlib
seaborn
sklearn
streamlit
Pillow